requirements.txt
asgiref==3.7.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
Django==4.2.1
django-ckeditor==6.5.1
django-classy-tags==4.0.0
django-contrib-comments==2.2.0
django-formtools==2.4.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-sekizai==4.1.0
django-treebeard==4.7
djangocms-admin-style==3.2.4
filebrowser-safe==1.1.1
grappelli-safe==1.1.1
idna==3.4
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tinycss2==1.1.1
typing-extensions==4.6.2
tzlocal==5.0.1
urllib3==2.0.2
webencodings==0.5.1
whitenoise==6.4.0
